loading the pages

+++ API_WMS/miebach/rest_api/views/inbound.py
@@ -6233,7 +6233,7 @@ def generate_grn(myDict, request, user, failed_qty_dict={}, passed_qty_dict={}, 
     if request.POST.get('grn_quantity', 0) == 'undefined':
         total_grn_qty = 0
     else:
-        total_grn_qty = int(request.POST.get('grn_quantity', 0))
+        total_grn_qty = float(request.POST.get('grn_quantity', 0))
     if request.POST.get('grn_total_amount', 0) == 'undefined':
         total_grn_value = 0
     else:
